PBO_SCALAR variable is undeclared in getOcMode
This is running on a desktop system with a Ryzen 1600.

The `if` from line 290 does not seem to set this variable, so when
running with the GUI and calling the getOcMode() function, the variable
is still undeclared.

NOTE: I do not know what the correct value should be to initialize the
PBO_SCALAR variable - I am using 0 because everything else there is
using 0.